+++ src/common/SignalVector.hpp
@@ -117,6 +117,12 @@ public:
         this->itemsChanged_();
     }
 
+    void clear()
+    {
+        assertInGuiThread();
+        this->items_.clear();
+    }
+
     const std::vector<T> &raw() const
     {
         assertInGuiThread();