+++ src/controllers/commands/CommandController.cpp
@@ -620,8 +620,6 @@ void CommandController::initialize(Settings &, Paths &paths)
             {
                 QString persistenceTime = persistence_list.at(1);
 
-                std::cout << persistenceTime.toStdString() << std::endl;
-
                 if (persistenceTime.endsWith("s"))
                 {
                     seconds =