+++ src/singletons/Settings.hpp
@@ -40,11 +40,15 @@ public:
     SignalVector<FilterRecordPtr> &filterRecords;
     SignalVector<Nickname> &nicknames;
     SignalVector<ModerationAction> &moderationActions;
+    SignalVector<QString> &separateLinksChannels;
 
     bool isHighlightedUser(const QString &username);
     bool isBlacklistedUser(const QString &username);
     bool isMutedChannel(const QString &channelName);
     bool toggleMutedChannel(const QString &channelName);
+    bool isSeparatedLinksChannel(const QString &channelName);
+    void addSeparatedLinkChannel(const QString &channelName);
+    void clearSeparatedLinkChannels();
 
 private:
     void mute(const QString &channelName);
@@ -178,6 +182,14 @@ public:
         "/behaviour/autocompletion/emoteCompletionWithColon", true};
     BoolSetting showUsernameCompletionMenu = {
         "/behaviour/autocompletion/showUsernameCompletionMenu", true};
+    BoolSetting enableHomiesCompletion = {
+        "/behaviour/autocompletion/enableHomiesCompletion", true};
+    BoolSetting enableFFZCompletion = {
+            "/behaviour/autocompletion/enableFFZCompletion", true};
+    BoolSetting enableBTTVCompletion = {
+            "/behaviour/autocompletion/enableBTTVCompletion", true};
+    BoolSetting enable7TVCompletion = {
+            "/behaviour/autocompletion/enable7TVCompletion", true};
 
     FloatSetting pauseOnHoverDuration = {"/behaviour/pauseOnHoverDuration", 0};
     EnumSetting<Qt::KeyboardModifier> pauseChatModifier = {
@@ -185,6 +197,8 @@ public:
     BoolSetting autorun = {"/behaviour/autorun", false};
     BoolSetting mentionUsersWithComma = {"/behaviour/mentionUsersWithComma",
                                          true};
+    BoolSetting joinSeparatedLinks = {"/behaviour/joinSeparatedLinks", false};
+    BoolSetting separateLinks = {"/behaviour/separateLinks", false};
 
     /// Commands
     BoolSetting allowCommandsAtEnd = {"/commands/allowCommandsAtEnd", false};
@@ -203,6 +217,12 @@ public:
         "/emotes/removeSpacesBetweenEmotes", false};
 
     BoolSetting enableHomiesGlobalEmotes = {"/emotes/enableHomiesGlobalEmotes",
+                                            true};
+    BoolSetting enable7TVGlobalEmotes = {"/emotes/enable7TVGlobalEmotes",
+                                            true};
+    BoolSetting enableFFZGlobalEmotes = {"/emotes/enableFFZGlobalEmotes",
+                                            true};
+    BoolSetting enableBTTVGlobalEmotes = {"/emotes/enableBTTVGlobalEmotes",
                                             true};
 
     /// Links