+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 ## Homies (#6)
+- Major: Added an option to separate and join URLs in selected channels
 - Minor: Added `Search in Google` option when text is selected
+- Minor: Added options to disable autocomplete for all global emotes providers
+- Minor: Added options to disable all global emotes providers
   
 - Minor: Added new search predicate to filter for messages matching a regex (#3282)
 - Minor: Subcategories in settings are now searchable. (#3157)